src/reviewservice/requirements.txt
Flask==2.0.1
requests==2.25.1
gunicorn==20.1.0
grpcio==1.59.2
grpcio-tools==1.59.2
protobuf>=4.0.0
pymongo==4.5.0